fixes some bugs and places placing order button in OrderSummary
