+++ frontend/src/components/constants/text.tsx
@@ -26,8 +26,10 @@ export const PAYMENT_METHOD_TEXT = "Payment Method"
 export const METHOD_TEXT = "Method"
 export const PAYMENT_METHOD_PAYBAL = "paypal"
 export const PAYMENT_METHOD_STRIPE = "stripe"
+export const PAYMENT_TEXT = "Payment"
 export const CONTINUE_TEXT = "Continue"
 export const CHOOSE_PAYMENT_METHOD_ERROR = "Choose a Payment Method first!"
+export const PAYMENT_METHOD_TITLE = "Method"
 export const PREVIEW_ORDER_TITLE = "Preview Order"
 export const SHIPPING_TITLE = "Shipping"
 export const NAME_TEXT = "Name"
@@ -38,3 +40,5 @@ export const ITEMS_TEXT = "Items"
 export const TAX_TEXT = "Tax"
 export const TOTAL_TEXT = "Total"
 export const PLACE_ORDER_TEXT = "Place Order"
+export const DELIVERED_TEXT = "Delivered"
+export const NOT_DELIVERED_TEXT = "Not Delivered"
