+++ frontend/src/components/orders/order-preview/PaymentUi.tsx
@@ -1,0 +1,34 @@
+import { Card, CardContent, Typography } from '@mui/material'
+import useOrdersContext from '../../../hooks/context/useOrdersContext';
+import { PAYMENT_METHOD_TITLE, PAYMENT_TEXT } from '../../constants/text';
+
+const PaymentUi = () => {
+    const { orderData } = useOrdersContext();
+
+    return (
+        <Card sx={ { minWidth: 275, width: "600px", marginBottom: "15px" } }>
+            <CardContent>
+                <Typography
+                    variant="h5"
+                    sx={ { marginBottom: "10px" } }
+                    fontWeight="fontWeightBold"
+                >
+                    { PAYMENT_TEXT }
+                </Typography>
+                <Typography variant="body1">
+                    <strong>{ PAYMENT_METHOD_TITLE }:</strong> { orderData?.paymentMethod }
+                </Typography>
+                <Typography
+                    sx={
+                        !orderData?.isPaid ? { color: "red" } : { color: "green" }
+                    }
+                    variant="body1"
+                >
+                    { orderData?.isPaid ? "Paid" : "Not Paid" }
+                </Typography>
+            </CardContent>
+        </Card>
+    )
+}
+
+export default PaymentUi
