+++ frontend/src/components/orders/order-preview/Title.tsx
@@ -1,0 +1,12 @@
+import { Typography } from '@mui/material'
+import { PREVIEW_ORDER_TITLE } from '../../constants/text'
+
+const Title = () => {
+    return (
+        <Typography id="order-preview-title" sx={ { marginBottom: "15px" } } variant="h3">
+            { PREVIEW_ORDER_TITLE }
+        </Typography>
+    )
+}
+
+export default Title
